+++ src/widgets/disk_table.rs
@@ -5,7 +5,7 @@ use tui::text::Text;
 
 use crate::{
     app::AppConfigFields,
-    canvas::canvas_colours::CanvasColours,
+    canvas::canvas_styling::CanvasColours,
     components::data_table::{
         ColumnHeader, DataTableColumn, DataTableProps, DataTableStyling, DataToCell, SortColumn,
         SortDataTable, SortDataTableProps, SortOrder, SortsRow,
